Update sample to use PerformanceWatcher
