+++ sample/ticktacktoe/script/measure_performance
@@ -12,9 +12,6 @@ sys.path.append(root)
 sys.path.append(src_path)
 sys.path.append(sample_path)
 
-import logging as log
-log.basicConfig(format='[%(levelname)s] %(message)s', level=log.INFO)
-
 from kyoka.algorithm.montecarlo.montecarlo import MonteCarlo
 from kyoka.algorithm.td_learning.sarsa import Sarsa
 from kyoka.algorithm.td_learning.q_learning import QLearning
@@ -30,7 +27,7 @@ from sample.ticktacktoe.ticktacktoe_table_value_function import TickTackToeTable
 from sample.ticktacktoe.ticktacktoe_helper import TickTackToeHelper
 from sample.ticktacktoe.ticktacktoe_manual_policy import TickTackToeManualPolicy
 from sample.ticktacktoe.ticktacktoe_perfect_policy import TickTackToePerfectPolicy
-from sample.ticktacktoe.ticktacktoe_performance_logger import TickTackToePerformanceLogger
+from sample.ticktacktoe.ticktacktoe_performance_watcher import TickTackToePerformanceWatcher
 
 SUPPORT_ALGORITHM = ["montecarlo", "sarsa", "qlearning", "sarsalambda", "qlambda", "minimax"]
 
@@ -59,14 +56,8 @@ RL_algo = {
     "qlambda": lambda :QLambda()
 }[algo]()
 
-callback = TickTackToePerformanceLogger()
-callback.set_performance_test_interval(TEST_INTERVAL)
-callback.set_is_first_player(is_first_player)
-callback.set_test_game_count(TEST_GAME_COUNT)
+callback = TickTackToePerformanceWatcher(TEST_INTERVAL, TEST_GAME_COUNT, is_first_player)
 
-log.info("start to measure performnce for %d episode" % TEST_LENGTH)
 RL_algo.setUp(domain, policy, value_func)
 RL_algo.run_gpi(TEST_LENGTH, callbacks=callback)
-log.info("finished to measure performnce for %d episode" % TEST_LENGTH)
-log.info("performance log = %s" % callback.game_log)
